+++ modules/project/src/main/java/com/falsepattern/zigbrains/project/steps/ui/BuildToolWindowFactory.java
@@ -1,0 +1,13 @@
+package com.falsepattern.zigbrains.project.steps.ui;
+
+import com.intellij.openapi.project.Project;
+import com.intellij.openapi.wm.ToolWindow;
+import com.intellij.openapi.wm.ToolWindowFactory;
+import org.jetbrains.annotations.NotNull;
+
+public class BuildToolWindowFactory implements ToolWindowFactory {
+    @Override
+    public void createToolWindowContent(@NotNull Project project, @NotNull ToolWindow toolWindow) {
+        BuildToolWindowContext.create(project, toolWindow);
+    }
+}